refactor(site/pages/AgentsPage): merge Limits and Usage into a single page
Combines the separate admin Limits and Usage settings sections into a
unified 'Usage & Limits' page. Admins can now see usage data alongside
limit controls, making it easier to set informed spend limits.

Changes:
- New UsageLimitsTab component composes DefaultLimitSection, users
  table, GroupLimitsSection, and UserOverridesSection into one view
- New UsageUsersSection extracts the paginated users table with search,
  drill-in, and effective spend limit column
- Sidebar merges Limits and Usage nav items into 'Usage & Limits'
- /agents/settings/limits redirects to the merged page
- Removes unused LimitsTab component
- Updates stories with limit config mocks and new header text
