--- site/src/pages/AgentsPage/components/LimitsTab/index.ts
@@ -1,1 +1,0 @@
-export { LimitsTab } from "./LimitsTab";
